+++ CHANGELOG.md
@@ -12,7 +12,10 @@
 * Add Admin note field to the Holdings field mapping profile: View (UIDATIMP-1122)
 * Add Admin note field to the Item field mapping profile: View (UIDATIMP-1123)
 * Display Authority information on Data import log page: View (UIDATIMP-1053)
+* Fix and Cover ViewAllLogs component with unit tests: Test (UIDATIMP-1151)
 * Add UI validation for some Invoice field mapping profile fields (UIDATIMP-1134)
+* Add summary at top of individual import job's log (UIDATIMP-1153)
+* For the Data Import View all page, change from Load more to Paginated (UIDATIMP-1147)
 * For long Data import logs, change from Load more to Paginated (UIDATIMP-1146)
 
 ### Bugs fixed: